fixed some make warnings while making SheepShaver

+++ SheepShaver/src/kpx_cpu/src/cpu/jit/dyngen.c
@@ -2079,7 +2079,7 @@ void gen_code(const char *name, const char *demangled_name,
 
                     if (val >= start_offset && val < start_offset + copy_size) {
                         n = strtol(p, NULL, 10);
-                        fprintf(outfile, "    label_offsets[%d] = %d + (code_ptr() - gen_code_buf);\n", n, val - start_offset);
+                        fprintf(outfile, "    label_offsets[%d] = %lu + (code_ptr() - gen_code_buf);\n", n, val - start_offset);
                     }
                 }
             }
@@ -2355,7 +2355,7 @@ void patch_relocations(FILE *outfile, const char *name, host_ulong size, host_ul
 					chaining: the offset of the instruction
 					needs to be stored */
 				fprintf(outfile, "    jmp_addr[%d] = code_ptr() + %d;\n",
-					n, rel->r_offset - start_offset);
+					n, (int)(rel->r_offset - start_offset);
 				continue;
 			}
 
@@ -2899,7 +2899,7 @@ int gen_file(FILE *outfile, int out_type)
 #ifdef CONFIG_FORMAT_MACH
 				fprintf(outfile, "DEFINE_CST(%s,0x%xL)\n\n", name, *((host_ulong *)(data + sym->st_value - data_sec_hdr->addr)));
 #else
-				fprintf(outfile, "DEFINE_CST(%s,0x%xL)\n\n", name, *((host_ulong *)(data + sym->st_value)));
+				fprintf(outfile, "DEFINE_CST(%s,0x%xL)\n\n", name, (unsigned int)(*((host_ulong *)(data + sym->st_value))));
 #endif
 			}
 			else if (strstart(name, OP_PREFIX "invoke", NULL)) {
